+++ packages/lib/src/transforms/index.ts
@@ -1,3 +1,4 @@
 export * from "./asMap"
 export * from "./asSet"
+export * from "./bigint"
 export * from "./date"