+++ packages/lib/src/action/pendingActions.ts
@@ -19,17 +19,24 @@ export function enqueuePendingAction(action: () => void): void {
   }
 }
 
+let pendingActionsRunning = false
+
 /**
  * @ignore
  */
-export function tryRunPendingActions(): boolean {
-  if (isActionRunning()) {
-    return false
+export function tryRunPendingActions(): void {
+  if (isActionRunning() || pendingActionsRunning) {
+    return
   }
 
-  while (pendingActions.length > 0) {
-    const nextAction = pendingActions.shift()!
-    nextAction()
+  pendingActionsRunning = true
+
+  try {
+    while (pendingActions.length > 0) {
+      const nextAction = pendingActions.shift()!
+      nextAction()
+    }
+  } finally {
+    pendingActionsRunning = false
   }
-  return true
 }